# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_pkg_utils.ErrorResponseSwag:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  backend_pkg_utils.SuccessResponseSwag:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  internal_proj_marketplace_handler.MessageResponse:
    properties:
      message:
        example: marketplace.operationSuccess
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_proj_marketplace_handler.ReindexStartedResponse:
    properties:
      message:
        example: marketplace.reindexStarted
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@svetu.rs
    name: API Support
    url: https://svetu.rs/support
  description: API для платформы Sve Tu - маркетплейс объявлений
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://svetu.rs/terms
  title: Sve Tu Marketplace API
  version: "1.0"
paths:
...
  /api/v1/admin/reindex-all-listings:
    post:
      consumes:
      - application/json
      description: Alternative endpoint to reindex all marketplace listings
      produces:
      - application/json
      responses:
        "200":
          description: Reindexing started
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.ReindexStartedResponse'
              type: object
        "401":
          description: marketplace.authRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "403":
          description: marketplace.adminRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.adminCheckError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Reindex all listings (alias)
      tags:
      - marketplace-admin-indexing
  /api/v1/admin/reindex-listings:
    post:
      consumes:
      - application/json
      description: Reindexes all marketplace listings in the search index
      produces:
      - application/json
      responses:
        "200":
          description: Reindexing started
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.MessageResponse'
              type: object
        "401":
          description: marketplace.authRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "403":
          description: marketplace.adminRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.adminCheckError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Reindex all listings
      tags:
      - marketplace-admin-indexing
  /api/v1/admin/reindex-listings-with-translations:
    post:
      consumes:
      - application/json
      description: Reindexes all marketplace listings with their translations in the
        search index
      produces:
      - application/json
      responses:
        "200":
          description: Reindexing with translations started
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.ReindexStartedResponse'
              type: object
        "401":
          description: marketplace.authRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "403":
          description: marketplace.adminRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.adminCheckError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Reindex all listings with translations
      tags:
      - marketplace-admin-indexing
  /api/v1/admin/reindex-ratings:
    post:
      consumes:
      - application/json
      description: Reindexes ratings and review counts for all marketplace listings
      produces:
      - application/json
      responses:
        "200":
          description: Rating reindexing started
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.ReindexStartedResponse'
              type: object
        "401":
          description: marketplace.authRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "403":
          description: marketplace.adminRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.adminCheckError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Reindex listing ratings
      tags:
      - marketplace-admin-indexing
...
securityDefinitions:
  Bearer:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"